+++ CHANGELOG.md
@@ -9,7 +9,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Validate ClinVar submission objects using the ClinVar API
 - Wrote tests for case and variant API endpoints
 - Create ClinVar submissions from Scout using the ClinVar API
-### Changed
+- Export Phenopacket for affected individual
+- Import Phenopacket from JSON file or Phenopacket API backend server
 - Use the new case name option for GENS requests
 ### Fixed
 - Fallback for empty alignment index for REViewer service
@@ -17,7 +18,6 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Avoid duplicate causatives and pinned variants
 - ClinVar clinical significance displays only the ACMG terms when user selects ACMG 2015 as assertion criteria
 - Spacing between icon and text on Beacon and MatchMaker links on case page sidebar
-
 
 
 ## [4.61.1]